rename intermediate_handlers module to enrichers
